Fix bugs in delete post funcionality

+++ src/controllers/newPostController.js
@@ -46,11 +46,13 @@ export async function newPost(_req, res) {
 }
 
 export async function deletePost(_req, res) {
-  const { postId } = res.locals.postId;
+  const { postId } = res.locals;
 
   try {
-    await newPostRepository.deletePostFromHashtagsPosts(postId);
-    await newPostRepository.deletePostById(postId);
+    const response = await newPostRepository.deletePostFromHashtagsPosts(
+      postId
+    );
+    if (response) await newPostRepository.deletePostById(postId);
     return res.status(202).send("Post deleted");
   } catch (err) {
     console.log(`Error controller: ${err}`);
